pyproject.toml
[project]
name = "echodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "SQLAlchemy>=2.0",
    "pydantic>=2.7",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
]
